// src/app/pages/home/models/enriched-data.ts
export interface EnrichedData {

    diff_totale_casi?: number;

    diff_percent_totale_casi?: number;

    diff_tamponi?: number;

    diff_percent_tamponi?: number;

    diff_deceduti?: number;

    diff_percent_deceduti?: number;

    diff_dimessi_guariti?: number;

    diff_percent_dimessi_guariti?: number;

    diff_nuovi_positivi?: number;

    diff_percent_nuovi_positivi?: number;

    diff_totale_positivi?: number;

    diff_percent_totale_positivi?: number;

    diff_totale_ospedalizzati?: number;

    diff_percent_totale_ospedalizzati?: number;

    diff_terapia_intensiva?: number;

    diff_percent_terapia_intensiva?: number;

    diff_ricoverati_con_sintomi?: number;

    diff_percent_ricoverati_con_sintomi?: number;
}